Move spec-oriented tests to *_spec_test.go files.
We started doing this and then stopped.
